Fix errors

+++ Ruby_Basics_part_1/perfect_weight.rb
@@ -2,7 +2,7 @@ puts "Как вас зовут?"
 first_name = gets.chomp
 
 puts "Какой у вас рост (см)?"
-height = gets.chomp.to_i
+height = gets.chomp.to_f
 
 if height - 110 < 0
   puts "Ваш вес уже оптимальный."
